refactor(frontend): organize components into feature subdirectories
Move components into command/, keyboard/, notification/,
preview/, and terminal/ subdirectories and update all
import paths accordingly.
