+++ frontend/src/App.vue
@@ -78,18 +78,18 @@
 
 <script setup lang="ts">
 import { ref, reactive, computed, watch, onMounted, onBeforeUnmount, nextTick } from 'vue'
-import TabBar from './components/TabBar.vue'
-import type { TabInfo } from './components/TabBar.vue'
-import TerminalPane from './components/TerminalPane.vue'
-import CommandPalette from './components/CommandPalette.vue'
-import type { Command } from './components/CommandPalette.vue'
-import MobileKeyboard from './components/MobileKeyboard.vue'
-import KbToggleButton from './components/KbToggleButton.vue'
+import TabBar from './components/terminal/TabBar.vue'
+import type { TabInfo } from './components/terminal/TabBar.vue'
+import TerminalPane from './components/terminal/TerminalPane.vue'
+import CommandPalette from './components/command/CommandPalette.vue'
+import type { Command } from './components/command/CommandPalette.vue'
+import MobileKeyboard from './components/keyboard/MobileKeyboard.vue'
+import KbToggleButton from './components/keyboard/KbToggleButton.vue'
 import SettingsPanel from './components/SettingsPanel.vue'
-import PreviewPanel from './components/PreviewPanel.vue'
-import CommandBookmarks from './components/CommandBookmarks.vue'
+import PreviewPanel from './components/preview/PreviewPanel.vue'
+import CommandBookmarks from './components/command/CommandBookmarks.vue'
 import ServerList from './components/ServerList.vue'
-import StatusBar from './components/StatusBar.vue'
+import StatusBar from './components/terminal/StatusBar.vue'
 import type { SyncServerMsg, SyncClientMsg } from './types/protocol'
 import { useSettings } from './composables/useSettings'
 import { getApiBase, wsUrlWithToken } from './composables/apiBase'
@@ -98,7 +98,7 @@ import { isTouchDevice } from './composables/useTerminal'
 import { useI18n } from './composables/useI18n'
 import { isWebPreviewInput } from './utils/previewRouting'
 import { initMonitorHistory } from './composables/useMonitorHistory'
-import NotificationPanel from './components/NotificationPanel.vue'
+import NotificationPanel from './components/notification/NotificationPanel.vue'
 import { useNotification } from './composables/useNotification'
 import { Settings, Bell, PanelRight, Plus, X, Star, AppWindow } from 'lucide-vue-next'
