+++ frontend/src/components/command/CommandPalette.vue
