+++ frontend/src/components/keyboard/SuggestionBar.vue
@@ -24,7 +24,7 @@
 
 <script setup lang="ts">
 import { ref } from 'vue'
-import type { SuggestionItem } from '../composables/useHistory'
+import type { SuggestionItem } from '../../composables/useHistory'
 
 defineProps<{
   suggestions: SuggestionItem[]
