+++ frontend/src/components/terminal/TabBar.vue
